Kwd 429 validate input fields protect against injection attacks (#890)
* working on improving input validation

* adds tests for new validations; adds validation to user controller routes

+++ src/controllers/user.ts
@@ -1,7 +1,7 @@
+import { SafeParseError, z, ZodError } from 'zod';
 import * as UserService from '../services/user';
 import * as UserUtils from '../services/user/utils';
 import * as output from '../services/output';
-import { verifyRequiredFields } from '../lib/requestData';
 import { ErrorTypes } from '../lib/constants';
 import CustomError, { asCustomError } from '../lib/customError';
 import { IRequestHandler } from '../types/request';
@@ -10,24 +10,28 @@ import * as UserTestIdentityService from '../services/user/testIdentities';
 import * as UserServiceTypes from '../services/user/types';
 import * as SupportTicketService from '../services/supportTicket';
 import { KWRateLimiterKeyPrefixes, setRateLimiterHeaders, unblockFromEmailLimiterOnSuccess } from '../middleware/rateLimiter';
+import { formatZodFieldErrors, getZodEnumSchemaFromTypescriptEnum, nameValidation, nanoIdValidation, optionalNameValidation, optionalObjectReferenceValidation, optionalUuidValidation, optionalZipCodeValidation } from '../lib/validation';
+import { DeleteRequestReason } from '../models/deleteAccountRequest';
 
 export const register: IRequestHandler<{}, {}, UserServiceTypes.IUserData> = async (req, res) => {
   try {
     const { body } = req;
-    const requiredFields = ['password', 'token', 'name'];
 
-    const { isValid, missingFields } = verifyRequiredFields(requiredFields, body);
-    if (!isValid) {
-      output.error(
-        req,
-        res,
-        new CustomError(
-          `Invalid input. Body requires the following fields: ${missingFields.join(', ')}.`,
-          ErrorTypes.INVALID_ARG,
-        ),
+    const registerUserSchema = z.object({
+      password: z.string(),
+      token: nanoIdValidation,
+      name: nameValidation,
+      promo: optionalObjectReferenceValidation,
+    });
+
+    const parsed = registerUserSchema.safeParse(req.body);
+    if (!parsed.success) {
+      const formattedError = formatZodFieldErrors(
+        ((parsed as SafeParseError<UserServiceTypes.IUserData>)?.error as ZodError)?.formErrors?.fieldErrors,
       );
-      return;
+      throw new CustomError(`${formattedError || 'Error parsing request'}`, ErrorTypes.INVALID_ARG);
     }
+
     const { password, name, token, promo } = body;
     const { user, authKey, groupCode } = await UserService.register({ password, name, token, promo });
     output.api(req, res, { user: UserUtils.getShareableUser(user), groupCode }, authKey);
@@ -40,6 +44,33 @@ export const login: IRequestHandler<{}, {}, UserServiceTypes.ILoginData> = async
   try {
     // TODO: limit failed attempts w/ https://github.com/animir/node-rate-limiter-flexible/wiki/Overall-example#minimal-protection-against-password-brute-force
     const { password, email, biometricSignature, fcmToken, deviceInfo } = req.body;
+
+    const loginSchema = z.object({
+      email: z.string().email(),
+      password: z.string(),
+      biometricSignature: z.string().optional(),
+      fcmToken: z.string().optional(),
+      deviceInfo: z
+        .object({
+          manufacturer: z.string().optional(),
+          bundleId: z.string().optional(),
+          deviceId: z.string().optional(),
+          apiLevel: z.string().optional(),
+          applicationName: z.string().optional(),
+          model: z.string().optional(),
+          buildNumber: z.string().optional(),
+        })
+        .optional(),
+    });
+
+    const parsed = loginSchema.safeParse(req.body);
+    if (!parsed.success) {
+      const formattedError = formatZodFieldErrors(
+        ((parsed as SafeParseError<UserServiceTypes.ILoginData>)?.error as ZodError)?.formErrors?.fieldErrors,
+      );
+      throw new CustomError(`${formattedError || 'Error parsing request'}`, ErrorTypes.INVALID_ARG);
+    }
+
     const { user, authKey } = await UserService.login(req, {
       biometricSignature,
       password,
@@ -57,6 +88,17 @@ export const login: IRequestHandler<{}, {}, UserServiceTypes.ILoginData> = async
 
 export const deleteAccountRequest: IRequestHandler<{}, {}, UserServiceTypes.IDeleteAccountRequest> = async (req, res) => {
   try {
+    const deleteAccountRequestSchema = z.object({
+      reason: getZodEnumSchemaFromTypescriptEnum(DeleteRequestReason),
+    });
+
+    const parsed = deleteAccountRequestSchema.safeParse(req.body);
+    if (!parsed.success) {
+      const formattedError = formatZodFieldErrors(
+        ((parsed as SafeParseError<UserServiceTypes.IDeleteAccountRequest>)?.error as ZodError)?.formErrors?.fieldErrors,
+      );
+      throw new CustomError(`${formattedError || 'Error parsing request'}`, ErrorTypes.INVALID_ARG);
+    }
     const response = await UserService.deleteAccountRequest(req);
     output.api(req, res, response);
   } catch (err) {
@@ -92,6 +134,26 @@ export const logout: IRequestHandler = async (req, res) => {
 
 export const updateProfile: IRequestHandler<{}, {}, UserServiceTypes.IUserData> = async (req, res) => {
   try {
+    const updateProfileSchema = z.object({
+      name: optionalNameValidation,
+      email: z.string().email().optional(),
+      zipcode: optionalZipCodeValidation,
+      integrations: z
+        .object({
+          marqeta: z.object({
+            userToken: optionalUuidValidation,
+          }).optional(),
+        })
+        .optional(),
+    });
+
+    const parsed = updateProfileSchema.safeParse(req.body);
+    if (!parsed.success) {
+      const formattedError = formatZodFieldErrors(
+        ((parsed as SafeParseError<UserServiceTypes.IUserData>)?.error as ZodError)?.formErrors?.fieldErrors,
+      );
+      throw new CustomError(`${formattedError || 'Error parsing request'}`, ErrorTypes.INVALID_ARG);
+    }
     const user = await UserService.updateProfile(req);
     output.api(req, res, UserUtils.getShareableUser(user));
   } catch (err) {
@@ -120,6 +182,17 @@ export const createPasswordResetToken: IRequestHandler<{}, {}, UserServiceTypes.
 
 export const verifyPasswordResetToken: IRequestHandler<{}, {}, UserServiceTypes.IVerifyTokenBody> = async (req, res) => {
   try {
+    const verifyPasswordResetTokenSchema = z.object({
+      token: nanoIdValidation,
+    });
+
+    const parsed = verifyPasswordResetTokenSchema.safeParse(req.body);
+    if (!parsed.success) {
+      const formattedError = formatZodFieldErrors(
+        ((parsed as SafeParseError<UserServiceTypes.IVerifyTokenBody>)?.error as ZodError)?.formErrors?.fieldErrors,
+      );
+      throw new CustomError(`${formattedError || 'Error parsing request'}`, ErrorTypes.INVALID_ARG);
+    }
     const data = await UserService.verifyPasswordResetToken(req);
     output.api(req, res, data);
   } catch (err) {
@@ -129,6 +202,17 @@ export const verifyPasswordResetToken: IRequestHandler<{}, {}, UserServiceTypes.
 
 export const checkIfEmailAlreadyInUse: IRequestHandler<{}, {}, UserServiceTypes.IEmail> = async (req, res) => {
   try {
+    const checkIfEmailAlreadyInUseSchema = z.object({
+      email: z.string().email(),
+    });
+
+    const parsed = checkIfEmailAlreadyInUseSchema.safeParse(req.body);
+    if (!parsed.success) {
+      const formattedError = formatZodFieldErrors(
+        ((parsed as SafeParseError<UserServiceTypes.IEmail>)?.error as ZodError)?.formErrors?.fieldErrors,
+      );
+      throw new CustomError(`${formattedError || 'Error parsing request'}`, ErrorTypes.INVALID_ARG);
+    }
     const data = await UserVerificationService.verifyUserDoesNotAlreadyExist(req);
     output.api(req, res, data);
   } catch (err) {
@@ -136,12 +220,23 @@ export const checkIfEmailAlreadyInUse: IRequestHandler<{}, {}, UserServiceTypes.
   }
 };
 
-export const resetPasswordFromToken: IRequestHandler<
-{},
-{},
-UserServiceTypes.ILoginData & UserServiceTypes.IUpdatePasswordBody
-> = async (req, res) => {
+export const resetPasswordFromToken: IRequestHandler<{}, {}, UserServiceTypes.ILoginData & UserServiceTypes.IUpdatePasswordBody> = async (
+  req,
+  res,
+) => {
   try {
+    const resetPasswordFromTokenSchema = z.object({
+      newPassword: z.string(),
+      token: nanoIdValidation,
+    });
+
+    const parsed = resetPasswordFromTokenSchema.safeParse(req.body);
+    if (!parsed.success) {
+      const formattedError = formatZodFieldErrors(
+        ((parsed as SafeParseError<UserServiceTypes.ILoginData & UserServiceTypes.IUpdatePasswordBody>)?.error as ZodError)?.formErrors?.fieldErrors,
+      );
+      throw new CustomError(`${formattedError || 'Error parsing request'}`, ErrorTypes.INVALID_ARG);
+    }
     const user = await UserService.resetPasswordFromToken(req);
     output.api(req, res, UserUtils.getShareableUser(user));
     await unblockFromEmailLimiterOnSuccess(req, res, KWRateLimiterKeyPrefixes.Login);
@@ -150,10 +245,7 @@ UserServiceTypes.ILoginData & UserServiceTypes.IUpdatePasswordBody
   }
 };
 
-export const resendEmailVerification: IRequestHandler<{}, {}, Partial<UserServiceTypes.IEmailVerificationData>> = async (
-  req,
-  res,
-) => {
+export const resendEmailVerification: IRequestHandler<{}, {}, Partial<UserServiceTypes.IEmailVerificationData>> = async (req, res) => {
   try {
     const data = await UserVerificationService.resendEmailVerification(req);
     output.api(req, res, data);
@@ -164,6 +256,18 @@ export const resendEmailVerification: IRequestHandler<{}, {}, Partial<UserServic
 
 export const verifyEmail: IRequestHandler<{}, {}, Partial<UserServiceTypes.IEmailVerificationData>> = async (req, res) => {
   try {
+    const verifyEmailSchema = z.object({
+      tokenValue: nanoIdValidation,
+    });
+
+    const parsed = verifyEmailSchema.safeParse(req.body);
+    if (!parsed.success) {
+      const formattedError = formatZodFieldErrors(
+        ((parsed as SafeParseError<UserServiceTypes.IEmailVerificationData>)?.error as ZodError)?.formErrors?.fieldErrors,
+      );
+      throw new CustomError(`${formattedError || 'Error parsing request'}`, ErrorTypes.INVALID_ARG);
+    }
+
     const data = await UserVerificationService.verifyEmail(req);
     output.api(req, res, data);
   } catch (err) {
